breakable repairboxes compatibility
breakable repairboxes is based on changing the attachment location for the original repairboxes to nothing, making them all go to 0,0,0

and manually inserting new repairboxes to script with. but this meant all new fabricators/recyclers would spawn with a hole where their repair box should be

this makes it so the change is done as a new kind of recycler/fabricator which all existing ones are converted to. but if someone makes a new one, it'll be fine, it just won't break when it's hit