big bug fix for breakable repair boxes
the linking method I used to link the new repair boxes with the recyclers/fabricators appears to only properly work with grav shafts and the like.

when it's used for recyclers/fabricators, it appears to work perfectly, but the machine reverts to its base state upon reloading the game

this overhauls the entire implementation, adding gametoken-based logic that makes sure they persistently behave as they should. this isn't extended to non-recycler/fabricator repair boxes as they already work fine